Fix Roz's multi-log closing problem, and get 0.9.9.6 locked in!

+++ Log System/RDLogCloser.m
@@ -52,7 +52,7 @@
 {
     NSIndexSet *logsToClose = [_rdTableView selectedRowIndexes];
     
-    unsigned nextToClose = [logsToClose firstIndex];
+    NSUInteger nextToClose = [logsToClose firstIndex];
 
     NSMutableArray *closers = [[NSMutableArray alloc] init];
